x11: initial implementation of from_existing_window
This adds the ability to create a glutin Window that wraps a pre-existing X
Window. The proximal motivation for this is to allow writing of XScreensaver
hacks in Rust, but it might also be useful for embedding 3D graphics in larger
applications with full process separation, etc.

This commit includes a bit of inline documentation, but no tests, and the
details of what if any WindowAttributes or pixel information should be applied
to the existing window aren't worked out.